requirements.txt
anyio==4.11.0 ; python_version >= "3.13" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.13" and python_version < "4.0"
cachecontrol==0.14.3 ; python_version >= "3.13" and python_version < "4.0"
cachetools==6.2.0 ; python_version >= "3.13" and python_version < "4.0"
certifi==2025.10.5 ; python_version >= "3.13" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.13" and platform_python_implementation != "PyPy" and python_version < "4.0"
charset-normalizer==3.4.3 ; python_version >= "3.13" and python_version < "4.0"
click==8.3.0 ; python_version >= "3.13" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4.0" and platform_system == "Windows"
cryptography==46.0.2 ; python_version >= "3.13" and python_version < "4.0"
firebase-admin==7.1.0 ; python_version >= "3.13" and python_version < "4.0"
flask==3.1.2 ; python_version >= "3.13" and python_version < "4.0"
google-api-core==2.26.0 ; python_version >= "3.13" and python_version < "4.0"
google-api-core[grpc]==2.26.0 ; python_version >= "3.13" and python_version < "4.0" and platform_python_implementation != "PyPy"
google-auth==2.41.1 ; python_version >= "3.13" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.13" and python_version < "4.0"
google-cloud-firestore==2.21.0 ; python_version >= "3.13" and python_version < "4.0" and platform_python_implementation != "PyPy"
google-cloud-storage==3.4.1 ; python_version >= "3.13" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.13" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.13" and python_version < "4.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.13" and python_version < "4.0"
grpcio-status==1.75.1 ; python_version >= "3.13" and platform_python_implementation != "PyPy" and python_version < "4.0"
grpcio==1.75.1 ; python_version >= "3.13" and platform_python_implementation != "PyPy" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.13" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.13" and python_version < "4.0"
h2==4.3.0 ; python_version >= "3.13" and python_version < "4.0"
hpack==4.1.0 ; python_version >= "3.13" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.13" and python_version < "4.0"
httpx[http2]==0.28.1 ; python_version >= "3.13" and python_version < "4.0"
hyperframe==6.1.0 ; python_version >= "3.13" and python_version < "4.0"
idna==3.10 ; python_version >= "3.13" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.13" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.13" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.13" and python_version < "4.0"
msgpack==1.1.2 ; python_version >= "3.13" and python_version < "4.0"
numpy==2.3.3 ; python_version >= "3.13" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.13" and python_version < "4.0"
pandas==2.3.3 ; python_version >= "3.13" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.13" and python_version < "4.0"
protobuf==6.32.1 ; python_version >= "3.13" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.13" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.13" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.13" and platform_python_implementation != "PyPy" and python_version < "4.0" and implementation_name != "PyPy"
pyjwt[crypto]==2.10.1 ; python_version >= "3.13" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.13" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.13" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.13" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.13" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.13" and python_version < "4"
six==1.17.0 ; python_version >= "3.13" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.13" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.13" and platform_python_implementation != "PyPy" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.13" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.13" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.13" and python_version < "4.0"
